user authentication

+++ controller/expense.js
@@ -17,7 +17,9 @@ const addexpense = async (req, res) => {
         if(isexpensevalid(expenseamount) || isexpensevalid(description) || isexpensevalid(category)){
             return res.status(400).json({succese: false, message: "Parameters missing"});
         }  
-        const expense = await Expense.create({expenseamount, description, category});
+        const userId = req.user.id;
+        const expense = await Expense.create({expenseamount, description, category, userId});
+        // const expense = await req.user.creatExpenses({expenseamount, description, category});
         return res.status(201).json({expense, succese: true});   
     } 
     catch (err) {
@@ -27,7 +29,7 @@ const addexpense = async (req, res) => {
 
 const getexpenses = async (req, res) => {
     try {
-        const expenses = await Expense.findAll();
+        const expenses = await Expense.findAll({where: {userId: req.user.id}});
         return res.status(200).json({expenses, succese: true});   
     } 
     catch (err) {
@@ -42,12 +44,16 @@ const deleteexpense = async (req, res) => {
         {
             return res.status(400).json({succese: false});
         }
-        await Expense.destroy({where: {id: expenseid}});
+        
+        const noOfRows = await Expense.destroy({where: {id: expenseid, userId: req.user.id}});
+        if(noOfRows === 0){
+            return res.status(404).json({succese: false, message: "Expense does not belongs to User"});
+        }
         return res.status(200).json({succese: true, message: "Deleted Successfully"});
         
     } 
     catch (err) {
-        return res.status(403).json({succese: true, message: "Failed"})
+        return res.status(403).json({succese: false, message: "Failed"})
     }
 }
